Improve and refactor HMI (incl Nextion display)
Implement the configuration of the HMI. The center and extreme positions
of the sticks can be teached via the menu on the Nextion display. These
settings are stored in the non-volatile memory (EEPROM) of the HMI
itself.

Improve high beam handling. The high beam is automatically switched of
when the low beam is off, similar to real cars.

Refactor the Teensy firmware, focusing on self-documenting code instead of
Doxygen comments.

Use 115200 baud for communication with Nextion HMI

+++ software/remote-control/vehicle-control-unit/src/steering-servo.cpp
@@ -35,7 +35,7 @@ namespace earth_rover_vcu {
             data[4] = configuration.maximum_pulse_width & 0x00ff;
             data[5] = (configuration.maximum_pulse_width & 0xff00) >> 8;
             data[6] = input_channel;
-            for(unsigned index = 7; index < size; ++index) {
+            for(uint16_t index = 7; index < size; ++index) {
                 data[index] = 0xff;
             }
             // Reset changed and save flags (we write this configuration to EEPROM, so changes are
